test: keep copied code off the system clipboard
